*: move "auth" to "internal/auth"
