+++ clientv3/main_test.go
@@ -22,8 +22,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/coreos/etcd/auth"
 	"github.com/coreos/etcd/integration"
+	"github.com/coreos/etcd/internal/auth"
 	"github.com/coreos/etcd/pkg/testutil"
 	"golang.org/x/crypto/bcrypt"
 )
