+++ internal/auth/range_perm_cache.go
@@ -15,7 +15,7 @@
 package auth
 
 import (
-	"github.com/coreos/etcd/auth/authpb"
+	"github.com/coreos/etcd/internal/auth/authpb"
 	"github.com/coreos/etcd/internal/mvcc/backend"
 	"github.com/coreos/etcd/pkg/adt"
 )
